update jenkins file

+++ Jenkinsfile
@@ -8,7 +8,15 @@ pipeline {
   stages {
     stage('Build') {
       steps {
+        script {
+                  if (env.BRANCH_NAME == 'master') {
         sh 'docker build -t localhost:5000/trade_frontend . && docker push localhost:5000/trade_frontend'
+                
+                  } else {
+        sh 'docker build -t localhost:5000/dev_trade_frontend . && docker push localhost:5000/dev_trade_frontend'
+                       
+                  }
+                }
       }
     }
     stage('Test') {
@@ -19,9 +27,28 @@ pipeline {
     }
     stage('Deploy') {
       steps {
-        sh 'docker stop trade_frontend'
-        sh 'docker rm trade_frontend'
-        sh 'docker run -d --name trade_frontend -p 3000:3000 localhost:5000/trade_frontend /bin/bash -c "npm run start"'
+        script{
+            if (env.BRANCH_NAME == 'master') {
+                      try {
+                        try {
+                            sh 'docker stop trade_frontend'
+                            sh 'docker rm trade_frontend'
+                        } catch (e) {
+                            echo 'container not running already'
+                        }
+                        sh 'docker run -d --name trade_frontend -p 3000:3000 localhost:5000/trade_frontend /bin/bash -c "npm run start"'
+                    
+                      } else {
+                        try {
+                            sh 'docker stop dev_trade_frontend'
+                            sh 'docker rm dev_trade_frontend'
+                        } catch (e) {
+                            echo 'container not running already'
+                        }
+                        sh 'docker run -d --name dev_trade_frontend -p 3100:3100 localhost:5000/dev_trade_frontend /bin/bash -c "npm run dev -- -p 3100"'
+                    
+                      }
+        }
       }
     }
   }
